🛡️ fix(config): decouple discovery auth from provider adapter
