+++ apps/www/src/content/docs/reference/provider-model-config.md
@@ -132,7 +132,7 @@ Example config documents:
 | `api_key` | `Option<RedactedString>` | `None` | Wrapped in `RedactedString`; redacted in `Debug`/`Display`. Wire format is a plain JSON string. Empty-string input deserializes as `None` so a stored key is preserved when the field is omitted on update |
 | `base_url` | `Option<String>` | `None` | Override base URL for proxies or self-hosted deployments. Empty-string input deserializes as `None` |
 | `timeout_secs` | `u64` | `300` | Request timeout in seconds |
-| `adapter_options` | `BTreeMap<String, Value>` | `{}` | Adapter-specific non-secret options. The OpenAI-compatible adapter recognizes `headers` (an object of string→string pairs added as default request headers). `model_discovery_schema` (`"openai"`/`"openai-compatible"` or `"gemini"`) opts a custom adapter into `/models` capability discovery using that schema (see Model capability sources). Unknown keys are accepted by the schema and ignored at build time. Secrets must use `api_key`; do not store credentials here |
+| `adapter_options` | `BTreeMap<String, Value>` | `{}` | Adapter-specific non-secret options. The OpenAI-compatible adapter recognizes `headers` (an object of string→string pairs added as default request headers). `model_discovery_schema` (`"openai"`/`"openai-compatible"` or `"gemini"`) opts a custom adapter into `/models` capability discovery using that schema (see Model capability sources). `model_discovery_auth` (`"bearer"`, `"x-goog-api-key"`, or `"none"`) overrides only the discovery request auth header; by default OpenAI-compatible discovery uses `Authorization: Bearer` and Gemini discovery uses `x-goog-api-key`. Unknown keys are accepted by the schema and ignored at build time. Secrets must use `api_key`; do not store credentials here |
 
 `ProviderSpec` deserialization ignores unknown top-level fields for stored-config
 compatibility. Config write and validate surfaces call `validate_provider_spec`
@@ -186,8 +186,11 @@ Provider discovery coverage is adapter-specific. Only adapters with a known
 `/models` schema are probed: `openai`/`openrouter` (OpenAI-compatible) and
 `gemini`/`google` (Gemini). An unknown or custom adapter is never silently
 treated as OpenAI-compatible; to discover a custom OpenAI-/Gemini-compatible
-gateway, opt in with `adapter_options.model_discovery_schema`. Gemini/Google
-discovery currently backfills token limits only; configure `modalities` and
+gateway, opt in with `adapter_options.model_discovery_schema`. Discovery auth is
+orthogonal to the adapter: it defaults from the declared discovery schema and can
+be overridden with `adapter_options.model_discovery_auth` when a gateway uses a
+different header convention. Gemini/Google discovery currently backfills token
+limits only; configure `modalities` and
 `knowledge_cutoff` explicitly when those fields should drive runtime guards or
 context injection. Vertex models may still receive static heuristic metadata,
 but Vertex provider discovery is not enabled unless a future adapter supplies a
